refactor: Move CollectionPage content to CollectionClient component for better organization
- fix netlify build error
